LayoutTests/http/tests/cookies/third-party-cookie-relaxing.html

<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script src="resources/cookie-utilities.js"></script>
<script>
var expectedMessage = null;
var message = null;
var expectedCookies = ["", "", "", "", "", "foo=awesomevalue", "foo=awesomevalue", "", "foo=awesomevalue", "bar=anotherawesomevalue; foo=awesomevalue"];
var currentFunction = 0;
var functions = new Array(
    startNewTest, allowAllCookies, deleteAllCookies, echoCookies, restrictCookies, setFooCookie, echoCookies,
    startNewTest, allowAllCookies, deleteAllCookies, echoCookies, restrictCookies, setFooAndBarCookies, echoCookies,
    startNewTest, allowAllCookies, deleteAllCookies, echoCookies, setFooCookie, echoCookies, restrictCookies, deleteCookies, echoCookies,
    startNewTest, allowAllCookies, deleteAllCookies, echoCookies, setFooCookie, echoCookies, restrictCookies, setFooAndBarCookies, echoCookies
);

function allowAllCookies()
{
    if (window.testRunner)
        testRunner.setAlwaysAcceptCookies(true);
    runNextTestOrFinish();
}

function restrictCookies()
{
    if (window.testRunner)
        testRunner.setAlwaysAcceptCookies(false);
    runNextTestOrFinish();
}

function deleteAllCookies()
{
    window.frames[0].postMessage("deleteCookiesInPopup", "*");
    deleteCookies();
}

function deleteCookies()
{
    executeCommandInFrame("deleteCookies");
    expectedMessage = "Deleted all cookies";
}

function echoCookies()
{
    window.frames[0].postMessage("echoCookiesInPopup", "*");
    expectedMessage = "Cookies:" + expectedCookies.shift();
}

function executeCommandInFrame(command)
{
    window.frames[0].postMessage("executeCommand " + command, "*");
}

function setFooCookie()
{
    executeCommandInFrame("setFooCookie");
    expectedMessage = "Set the foo cookie";
}

function setFooAndBarCookies()
{
    executeCommandInFrame("setFooAndBarCookie");
    expectedMessage = "Set the foo and bar cookies";
}

function startNewTest()
{
    runNextTestOrFinish();
}

function onMessage(event)
{
    message = event.data;
    shouldBeEqualToString('message', expectedMessage);
    runNextTestOrFinish();
}

async function startTest()
{
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }

    await resetCookies();
    runNextTestOrFinish();
}

async function runNextTestOrFinish()
{
    if (!currentFunction)
        window.addEventListener("message", onMessage);

    if (currentFunction >= functions.length) {
        // resetCookies() will add message event listener, so remove the existing one.
        window.removeEventListener("message", onMessage);

        await resetCookies();
        if (window.testRunner)
            testRunner.notifyDone();
        return;
    }

    var functionToRun = currentFunction++;
    functions[functionToRun]();
}

</script>
</head>
<body onload="startTest()">
<iframe id='testFrame' src="http://localhost:8000/cookies/resources/third-party-cookie-relaxing-iframe.html"></iframe>
</body>
</html>
