LayoutTests/http/tests/cookies/third-party-cookie-relaxing-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS message is "Deleted all cookies"
PASS message is "Cookies:"
PASS message is "Set the foo cookie"
PASS message is "Cookies:"
PASS message is "Deleted all cookies"
PASS message is "Cookies:"
PASS message is "Set the foo and bar cookies"
PASS message is "Cookies:"
PASS message is "Deleted all cookies"
PASS message is "Cookies:"
PASS message is "Set the foo cookie"
FAIL message should be Cookies:foo=awesomevalue. Was Cookies:.
PASS message is "Deleted all cookies"
FAIL message should be Cookies:foo=awesomevalue. Was Cookies:.
PASS message is "Deleted all cookies"
PASS message is "Cookies:"
PASS message is "Set the foo cookie"
FAIL message should be Cookies:foo=awesomevalue. Was Cookies:.
PASS message is "Set the foo and bar cookies"
FAIL message should be Cookies:bar=anotherawesomevalue; foo=awesomevalue. Was Cookies:.
